Moved Gatekeeper User model methods to trait
